* master: (97 commits)
  Grouped all changes as fixes
  fixed typos in release notes
  Add vpgl to libraries for VXL tests
  fixed comment copy/paste error
  added release note
  decrease default detection threshold
  switch default configuration to use non-upright descriptors
  switch default configs to use keyframe loop closure
  explicit cast to avoid MSVC warning
  explicit cast to avoid MSVC warning
  convert OpenCV type integer into a string
  update Travis CI to use VXL master
  fixed signed/unsigned comparison warnings
  Add separate range and automatic values to colors
  Added release note.
  Fixed homogenous point initialisation in UnprojectPoint.
  Depth map UI cleanup, part II
  Don't crash on depth map load errors
  Use GetPointer consistently
  Avoid hard-coded-in-place strings
  ...

+++ maptk/plugins/core/close_loops_keyframe.cxx
@@ -363,7 +363,8 @@ close_loops_keyframe
                            << num_linked << " joined tracks");
   }
   // divide by number of matched frames to get the average
-  d_->frame_matches[frame_number] /= (frame_number - last_frame);
+  d_->frame_matches[frame_number] /=
+    static_cast<unsigned int>(frame_number - last_frame);
 
   // stitch with all previous keyframes
   for(auto kitr = d_->keyframes.rbegin(); kitr != d_->keyframes.rend(); ++kitr)
